+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Updated `checks` to `3.2.0`, `cartridge` to `2.7.8` and `metrics` to `0.16.0` in application template.
 
+- Set `compat.fiber_slice_default` to `new` by default in application template.
+
 ## [2.12.4] - 2022-12-31
 
 ### Changed