+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Update `checks` to `3.3.0`, `cartridge` to `2.8.1`, `metrics` to `1.0.0` and
+  `cartridge-metrics-role` to `0.1.1` in application template.
+
 ## [2.12.7] - 2023-08-01
 
 ### Changed