+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.12.9] - 2023-10-19
+
+### Changed
+
+- Updated the go version used to build the binary, that fixes known security issues.
+
 ## [2.12.8] - 2023-08-24
 
 ### Changed